Course.java
Courses are defined courses for students that admins and professors can define. Each course has its own fields. Also, in the local database, each course has two files. One for the students in that course and their grades, and one for the rest of the course details (professor name - number of units, etc.)